ECA-4630: Avoid some unnecessary Integer.valueOf calls for hashmap.remove()
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@31356 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/ejbca-common/src/org/ejbca/core/model/ra/raadmin/EndEntityProfile.java
@@ -408,7 +408,7 @@ public class EndEntityProfile extends UpgradeableDataHashMap implements Serializ
     	    // validation should be a map of a validator class name (excluding package name) and a validator-specific object.
     	    data.put(FIELDBOUNDRARY_VALIDATION + offset, validation);
     	} else {
-    	    data.remove(Integer.valueOf(FIELDBOUNDRARY_VALIDATION + offset)); // must use Integer.valueOf to avoid calling remove(index)
+    	    data.remove(FIELDBOUNDRARY_VALIDATION + offset);
     	}
     	if (DnComponents.isDnProfileField(parameterName)) {
     		@SuppressWarnings("unchecked")
@@ -472,10 +472,10 @@ public class EndEntityProfile extends UpgradeableDataHashMap implements Serializ
     			fieldorder.remove(value);
     		}
     		// Remove last element of the type from hashmap
-    		data.remove(Integer.valueOf(FIELDBOUNDRARY_VALUE + (NUMBERBOUNDRARY*(size-1)) + parameter));
-    		data.remove(Integer.valueOf(FIELDBOUNDRARY_USE + (NUMBERBOUNDRARY*(size-1)) + parameter));
-    		data.remove(Integer.valueOf(FIELDBOUNDRARY_ISREQUIRED + (NUMBERBOUNDRARY*(size-1)) + parameter));
-    		data.remove(Integer.valueOf(FIELDBOUNDRARY_MODIFYABLE + (NUMBERBOUNDRARY*(size-1)) + parameter));
+    		data.remove(FIELDBOUNDRARY_VALUE + (NUMBERBOUNDRARY*(size-1)) + parameter);
+    		data.remove(FIELDBOUNDRARY_USE + (NUMBERBOUNDRARY*(size-1)) + parameter);
+    		data.remove(FIELDBOUNDRARY_ISREQUIRED + (NUMBERBOUNDRARY*(size-1)) + parameter);
+    		data.remove(FIELDBOUNDRARY_MODIFYABLE + (NUMBERBOUNDRARY*(size-1)) + parameter);
     		decrementFieldnumber(parameter);
     	}
     }
@@ -555,7 +555,7 @@ public class EndEntityProfile extends UpgradeableDataHashMap implements Serializ
         if (validation != null) {
             data.put(paramNum, new LinkedHashMap<>(validation));
         } else {
-            data.remove(paramNum); // must use Integer for remove method
+            data.remove(paramNum);
         }
     }
